+++ Source/Util/util.h
@@ -9,6 +9,6 @@ enum LogMessage {
 
 void exitOnError(char* err, uint uxerr);
 void ulog(LogMessage level, char *message);
-void ulogFormat(LogMessage level, uint8_t size, char *message, const char * format, ... );
+void ulogFormat(LogMessage level, uint64_t size, char *message, const char * format, ... );
 
 #endif /* UTIL_DOT_H */
